refactor: remove common/tlsx/full isolation, inline cfssl into common/tlsx
cfssl compiles under go 1.11 so the go mod isolation is unnecessary.
Merge cfssl revocation check directly into common/tlsx, removing the
RegisterRevokeCheck indirection and the common/tlsx/full sub-module.

- Inline cfssl/revoke.VerifyCertificate into tlsx.IsRevoked
- Remove RevokeCheckFunc, RegisterRevokeCheck, HasRevokeCheck
- Delete common/tlsx/full/ (go.mod, revoke.go, revoke_test.go)
- Remove compile-time "no revocation backend" guard from SSL protocol
- Remove common/tlsx/full CI step
- Add cfssl as direct dependency in root go.mod
