+++ .github/workflows/ci.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Set up Go
         uses: actions/setup-go@v5
         with:
-          # operators/full and common/tlsx/full submodules require go 1.24.
+          # operators/full submodule requires go 1.24.
           go-version: '1.24'
 
       - name: Build
@@ -27,8 +27,8 @@ jobs:
 
       # Whole main module (convert, common/dsl, common/tlsx, operators,
       # protocols/{http,ssl}, harness, templates ...). `./...` does not descend
-      # into operators/full or common/tlsx/full — those are separate modules
-      # tested in their own steps below.
+      # into operators/full — that is a separate module tested in its own step
+      # below.
       - name: Test
         run: go test ./... -v -count=1 -timeout=180s
 
@@ -42,16 +42,6 @@ jobs:
       # the registration surface main relies on doesn't drift.
       - name: Test operators/full submodule
         working-directory: operators/full
-        run: |
-          go build ./...
-          go vet ./...
-          go test ./... -v -count=1 -timeout=120s
-
-      # common/tlsx/full is the opt-in cfssl-based certificate revocation check
-      # submodule. Like operators/full it lives in its own Go module so the main
-      # neutron module keeps a stdlib-only TLS path — run its tests separately.
-      - name: Test common/tlsx/full submodule
-        working-directory: common/tlsx/full
         run: |
           go build ./...
           go vet ./...
