+++ go.mod
@@ -3,7 +3,7 @@ module github.com/chainreactors/neutron
 go 1.11
 
 require (
-	github.com/Knetic/govaluate v3.0.0+incompatible
+	github.com/Knetic/govaluate v3.0.1-0.20171022003610-9aa49832a739+incompatible
 	github.com/chainreactors/logs v0.0.0-20260508055944-c678762ed15c
 	github.com/chainreactors/words v0.0.0-20260520145736-270600e60fb4
 	github.com/hashicorp/go-version v1.6.0
@@ -12,7 +12,8 @@ require (
 )
 
 require (
-	github.com/davecgh/go-spew v1.1.1
+	github.com/cloudflare/cfssl v1.6.5
+	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc
 	github.com/invopop/jsonschema v0.13.0
 	gopkg.in/yaml.v3 v3.0.1
 )
